+++ CMakeLists.txt
@@ -7,7 +7,7 @@ cmake_minimum_required(VERSION 3.1)
 project(RedisClient)
 set(VERSION_MAJOR 1 CACHE STRING "Project major version number.")
 set(VERSION_MINOR 0 CACHE STRING "Project minor version number.")
-set(VERSION_PATCH 0 CACHE STRING "Project patch version number.")
+set(VERSION_PATCH 1 CACHE STRING "Project patch version number.")
 set(VERSION ${VERSION_MAJOR}.${VERSION_MINOR}.${VERSION_PATCH})
 mark_as_advanced(VERSION_MAJOR VERSION_MINOR VERSION_PATCH VERSION)
