+++ src/redisclient/impl/pipeline.h
@@ -7,8 +7,8 @@
 
 #include <deque>
 
-#include "redisclient/redisbuffer.h"
-#include "redisclient/config.h"
+#include "redisclient/impl/redisbuffer.h"
+#include "redisclient/impl/config.h"
 
 namespace redisclient
 {
