+++ src/redisclient/impl/redisbuffer.h
@@ -12,7 +12,7 @@
 #include <string>
 #include <vector>
 
-#include "redisclient/config.h"
+#include "redisclient/impl/config.h"
 
 namespace redisclient {
