+++ src/redisclient/redisclient.h
@@ -1,13 +1,13 @@
 #pragma once
 #ifndef REDISCLIENT_REDISCLIENT_H
 #define REDISCLIENT_REDISCLIENT_H
-#include "redisclient/version.h"
-#include "redisclient/redisbuffer.h"
-#include "redisclient/redisvalue.h"
-#include "redisclient/redisparser.h"
+#include "redisclient/impl/version.h"
+#include "redisclient/impl/redisbuffer.h"
+#include "redisclient/impl/redisvalue.h"
+#include "redisclient/impl/redisparser.h"
 #include "redisclient/impl/redisclientimpl.h"
 #include "redisclient/impl/throwerror.h"
-#include "redisclient/redisasyncclient.h"
-#include "redisclient/redissyncclient.h"
-#include "redisclient/pipeline.h"
+#include "redisclient/impl/redisasyncclient.h"
+#include "redisclient/impl/redissyncclient.h"
+#include "redisclient/impl/pipeline.h"
 #endif
