+++ src/redisclient/version.h
@@ -6,6 +6,6 @@
 #ifndef REDISCLIENT_VERSION_H
 #define REDISCLIENT_VERSION_H
 
-#define REDIS_CLIENT_VERSION 600 // 0.6.0
+#define REDIS_CLIENT_VERSION 1000 // 1.0.1
 
 #endif // REDISCLIENT_VERSION_H
